+++ runtime/src/main/scala/vork/internal/document/DocumentBuilder.scala
@@ -72,7 +72,7 @@ trait DocumentBuilder {
       try {
         Console.withOut(myPs) {
           Console.withErr(myPs) {
-           app()
+            app()
           }
         }
       } catch {
