+++ vork/src/main/scala/vork/internal/markdown/SectionInput.scala
@@ -1,0 +1,6 @@
+package vork.internal.markdown
+
+import scala.meta.Source
+import scala.meta.inputs.Input
+
+case class SectionInput(input: Input, source: Source, mod: Modifier)
